refactor(server): extract zernio subsystem into 2 router modules
Eleventh route-group extraction — the zernio subsystem (10 of its 14
routes), split across 2 router files per the chosen structure (the route
guard scans one express.Router() var per file, so two routers can't share
a file).

- src/server/routes/zernio.js -> app.use('/api/zernio', zernioRouter)
  3 core routes (connect/:platform [open OAuth init], disconnect, connect).
  MIXED auth -> per-route. Deps: pool, requireAuth/verifyBrandAccess,
  callZernio/getOrCreateZernioProfile (zernio.js).
- src/server/routes/zernio-admin.js -> app.use('/api/admin/zernio', ...)
  7 dev/strategy-only test probes, each gated by zernioGuard. Deps: pool,
  zernioGuard/callZernio/getOrCreateZernioProfile/zernioPublish (zernio.js).

LEFT INLINE (deliberate, per scoping):
- /auth/zernio/callback + /integrations/zernio/callback — 2 OAuth callbacks
  at odd single-route prefixes with different logic (legacy vs newer flow).
- /api/admin/backfill-{facebook,linkedin}-zernio-ids — admin-lane one-shots
  under /api/admin (can't capture without an /api/admin router).

All zernio HTTP helpers were already in zernio.js — zero new coupling.
Handler bodies verbatim; only registration lines changed.

Test plan:
- node --check server.js + both modules -> OK
- npm run lint (no-undef) -> clean
- both modules boot-load (import()) -> resolve
- route-inventory guard -> 213 byte-identical (14 zernio paths total:
  10 reconstructed via the 2 mounts + 4 still inline)
- vitest -> 93/93

Rollback: revert; verbatim multi-span move into 2 modules.
